Onboarding: created expandable_buttons_section
