+++ lib/src/modules/onboarding/widgets/page/page.dart
@@ -1,1 +1,2 @@
 export 'onboarding_religion_preference_page.dart';
+export 'onboarding_v2.dart';
