+++ lib/src/modules/onboarding/widgets/widgets.dart
@@ -1,1 +1,3 @@
+export 'atoms/atoms.dart';
+export 'molecules/molecules.dart';
 export 'page/page.dart';
